:wrench: Fix hardcoded path and modularize model_setup
